requirements.txt
# Core
sqlalchemy==2.0.32
pymysql==1.1.1
python-dotenv==1.0.1
pydantic==2.8.2

# LLM (Gemini via LangChain)
langchain==0.2.15
langchain-google-genai==1.0.7
google-generativeai==0.7.2

# MCP server
fastmcp==0.4.1

# OCR stack
pytesseract==0.3.13
pillow==10.4.0
pdf2image==1.17.0

streamlit
